requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysqlclient==2.1.0
Werkzeug==2.0.2
